// kservices_ts_api/src/api/domain/entities/v2/ivr.entity.ts
export type updateS3Recording = {
  sme_id: number;
  in_flag: number;
  in_call_id: string;
  in_agent_id: number;
  in_customer_ani: string;
  in_duration: number;
  in_filename: string;
  in_status: number;
  in_flag_tbl: string;
  in_ip: string;
  insertDateTime: string;
  in_file: string;
  session_id: string;
  call_direction: string;
  crm_url: string;
  komm_url: string;
};

export type updateS3RecordingStatus = {
  sme_id: number;
  session_id: string;
  call_direction: string;
  getCurrentDate: string;
};